* Mehdify
Added algorithm which finds next larger number by scanning from right to
left. There might be slightly faster implementations, but this code
should be very readable and understandable.

I choosed to do it the oldschool way since the module was a function
already. This could be done with a ES6 class and usage of forEach/map
instead of for loops.